Unified providers into a single trait
