+++ src/config.rs
@@ -1,7 +1,9 @@
 pub const APP_ID: &str = "dev.edfloreshz.Done";
 pub const GETTEXT_PACKAGE: &str = "done";
 pub const LOCALEDIR: &str = "/usr/local/share/locale";
+#[allow(dead_code)]
 pub const PKGDATADIR: &str = "/usr/local/share/done";
 pub const PROFILE: &str = "";
+#[allow(dead_code)]
 pub const RESOURCES_FILE: &str = concat!("/usr/local/share/done", "/resources.gresource");
 pub const VERSION: &str = "0.1.1";
