+++ src/data/traits/provider.rs
@@ -1,8 +1,6 @@
 use std::fmt::Debug;
 
 use anyhow::Result;
-use diesel::SqliteConnection;
-use relm4::factory::DynamicIndex;
 use serde::{Deserialize, Serialize};
 
 use crate::data::models::generic::lists::GenericList;
@@ -10,49 +8,43 @@ use crate::data::models::generic::tasks::GenericTask;
 use crate::gtk;
 
 pub trait Provider: Debug {
-	/// The unique identifier of the `TaskProvider`.
+	/// Getters
+	///
+	/// The unique identifier of the provider.
 	fn get_id(&self) -> &str;
-	/// The user-visible name of the `TaskProvider`.
+	/// The user-visible name of the provider.
 	fn get_name(&self) -> &str;
-	/// The type of the `TaskProvider`.
+	/// The type of the provider.
 	fn get_provider_type(&self) -> ProviderType;
-	/// The description of the `TaskProvider`, e.g. the account user of a GNOME Online Accounts' account
+	/// The description of the provider, e.g. the account user of a GNOME Online Accounts' account
 	fn get_description(&self) -> &str;
-	/// Whether the `TaskProvider` is enabled.
-	fn get_enabled(&self) -> bool;
+	/// Whether the provider is enabled.
+	fn is_enabled(&self) -> bool;
+	/// Gets the icon name of the provider.
+	fn get_icon_name(&self) -> &str;
+	/// Gets the icon of the provider.
+	fn get_icon(&self) -> gtk::Image;
+
+	/// # Setters
+	///
 	/// Sets the provider as enabled.
 	fn set_enabled(&mut self);
+	/// Sets the provider as disabled.
+	fn set_disabled(&mut self);
+
+	/// Methods
+	///
+	/// Creates a new instance of the provider.
+	fn new() -> Self
+	where
+		Self: Sized;
 	/// Asks the provider to refresh. Online providers may want to
 	/// synchronize tasks and task lists, credentials, etc, when this
 	/// is called.
-	fn refresh(&self);
-	fn get_icon_name(&self) -> String;
-	fn get_icon(&self) -> gtk::Image;
-}
+	fn refresh(&self) -> Result<()>;
 
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub enum ProviderType {
-	Inbox,
-	Today,
-	Next7Days,
-	All,
-	Local,
-}
-
-pub trait Service: Debug + Send {
-	fn init() -> Self
-	where
-		Self: Sized;
-	fn establish_connection(&self) -> Result<SqliteConnection>;
-	// Fetch tasks from the provider and save them to local storage.
-	fn refresh_tasks(&mut self) -> Result<()>;
-	fn refresh_lists(&mut self) -> Result<()>;
-
-	fn get_provider(&self) -> Box<dyn Provider>;
-	fn get_tasks(&self) -> Vec<GenericTask>;
-	fn get_task_lists(&self) -> Vec<GenericList>;
-
-	// Tasks
+	/// Tasks
+	///
 	/// This method should return the list of tasks in a list.
 	fn read_tasks_from_list(&self, id: &str) -> Result<Vec<GenericTask>>;
 	/// This method should return the information about a task.
@@ -68,19 +60,28 @@ pub trait Service: Debug + Send {
 	/// This method should remove an existing task.
 	fn remove_task(&mut self, task_id: &str) -> Result<()>;
 
-	// Task Lists
+	/// Task Lists
+	///
 	/// This method should return the lists from a provider.
 	fn read_task_lists(&self) -> Result<Vec<GenericList>>;
 	/// This method should create a new list for a provider.
 	fn create_task_list(
 		&mut self,
-		provider: &str,
+		list_provider: &str,
 		name: &str,
 		icon: &str,
 	) -> Result<GenericList>;
 	/// This method should update an existing list for a provider.
-	fn update_task_list(&mut self, index: DynamicIndex, name: &str)
-		-> Result<()>;
+	fn update_task_list(&mut self, list: GenericList, name: &str) -> Result<()>;
 	/// This method should remove a list from a provider.
-	fn remove_task_list(&mut self, index: DynamicIndex) -> Result<()>;
+	fn remove_task_list(&mut self, list: GenericList) -> Result<()>;
+}
+
+#[derive(Debug, Copy, Serialize, Deserialize, Clone)]
+pub enum ProviderType {
+	Inbox,
+	Today,
+	Next7Days,
+	All,
+	Local,
 }
